Add verification of possible bindings in run/srun scripts
The ckains tool requires that explicit bindings are bindable,
hence we do a check before adding default bindings to run/srun.
A dir bin must exist on the source side and either not exist
or be a directory on the target side.
A file bind must exist on the source side and either not exist
or be a regular file (i.e. not a link) on the target side.

Add back /tmp as default bindings as it may cause issue with
PRoot for prograzms that assume a short path for the tmp dir
such as when creating unix domain sockets in the tmp dir.