Add a retry policy for downloads and updates
When dowloads or packages updates fail with a server side
HTTP error code in the 500-504 range, re-execute the operation.
The retry is done with an increasing delay up to a reasonable
delay and retry count.

This is mainly motivated to make zoostrap more robust to HTTP
proxies or download sites that may return a 503 service unavailable
code from time to time.