Add support for Ubuntu point releases from cdimage.ubuntu.com
Add support for Ubuntu point releases 16.04.1-5 and 18.04.1.

Use cdimage.ubuntu.com as source as it seems that the site
is now maintained again by Canonical and references the
the ubuntu point releases.

Note that sha1sum of tarballs are checked, though it is unclear
whether the released files on cdimage.ubuntu.com are supposed to
be immutables, this may have to be re-assessed.